+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,1 @@
-null
+- DependencyFailedException shows the root module that failed if it's part of a chain of dependent modules